Starkit V1.0
Starkit V1.0 is a cutting-edge wireless kit designed for seamless connectivity and versatility. Leveraging the power of ESP32-C6 and SIM7600E-H wireless modules, this kit brings a host of features to the table.

Key Kit Features:
- Wi-Fi 6 in 2.4 GHz Band, Bluetooth 5 (LE), Thread/Zigbee
- LTE 4G for high-speed data connectivity
- GNSS for precise location tracking
- Embedded Microphone and Speaker for cellular communication
- NanoSIM slot for SIM card integration
- 2 USB C connectors for power and debugging

With a total of 5 antenna slots, Starkit V1.0 provides great flexibility where users can seamlessly switch between and SMA antenna and a 2.4GHz chip antenna for the ESP32-C6 connectivity. There are 2 SMAs for the Main and AUX GSM antennas and finally there is a 1575.42MHz Patch Antenna for GNSS.

Starkit V1.0 is crafted with developers in mind. Whether you're working on IoT, connectivity projects, or exploring the boundaries of wireless technology, this kit enables developers to bring their ideas to life.